Money formatting
Now at least it puts the commas in the stream.  I don't know why it
won't put them in the outputFile, but it's probably better that it
doesn't anyway.  The put_money didn't work well at all.  For some
reason, it cuts off the pennies and uses the last 2 dollars as if
they were pennies.  That's why I'm not really trying to format the
money currently.